+++ CHANGELOG.md
@@ -64,7 +64,7 @@
   - `bract.core.util/thrown->val`
   - `bract.core.util/after`
   - `bract.core.util/doafter`
-- [Todo] Wrap fn returned by `bract.core.keydef/resolve-context` to ignore event-logging exceptions
+- Wrap fn returned by `bract.core.keydef/resolve-event-logger` to ignore event-logging exceptions
 
 
 ### 0.6.2-alpha4 / 2021-February-13